+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 0.4.2 - 2017-05-25
+### Refactored
+- some functions names
+
+### Added
+- exercise 4.4
+- gauss doolittle method
+
 ## 0.4.1 - 2017-05-24
 ### Fixed
 - L creation on Gauss solvers